docs/coverage/node/html/test/wpt/xhr/setrequestheader-header-forbidden.js.gcov.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">

<html lang="en">

<head>
  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
  <title>LCOV - lcov.info - test/wpt/xhr/setrequestheader-header-forbidden.js</title>
  <link rel="stylesheet" type="text/css" href="../../../gcov.css">
</head>

<body>

  <table width="100%" border=0 cellspacing=0 cellpadding=0>
    <tr><td class="title">LCOV - code coverage report</td></tr>
    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>

    <tr>
      <td width="100%">
        <table cellpadding=1 border=0 width="100%">
          <tr>
            <td width="10%" class="headerItem">Current view:</td>
            <td width="35%" class="headerValue"><a href="../../../index.html">top level</a> - <a href="index.html">test/wpt/xhr</a> - setrequestheader-header-forbidden.js<span style="font-size: 80%;"> (source / <a href="setrequestheader-header-forbidden.js.func-sort-c.html">functions</a>)</span></td>
            <td width="5%"></td>
            <td width="15%"></td>
            <td width="10%" class="headerCovTableHead">Hit</td>
            <td width="10%" class="headerCovTableHead">Total</td>
            <td width="15%" class="headerCovTableHead">Coverage</td>
          </tr>
          <tr>
            <td class="headerItem">Test:</td>
            <td class="headerValue">lcov.info</td>
            <td></td>
            <td class="headerItem">Lines:</td>
            <td class="headerCovTableEntry">194</td>
            <td class="headerCovTableEntry">194</td>
            <td class="headerCovTableEntryHi">100.0 %</td>
          </tr>
          <tr>
            <td class="headerItem">Date:</td>
            <td class="headerValue">2024-12-07 00:20:21</td>
            <td></td>
            <td class="headerItem">Functions:</td>
            <td class="headerCovTableEntry">1</td>
            <td class="headerCovTableEntry">1</td>
            <td class="headerCovTableEntryHi">100.0 %</td>
          </tr>
          <tr>
            <td></td>
            <td></td>
            <td></td>
            <td class="headerItem">Branches:</td>
            <td class="headerCovTableEntry">6</td>
            <td class="headerCovTableEntry">6</td>
            <td class="headerCovTableEntryHi">100.0 %</td>
          </tr>
          <tr><td><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
        </table>
      </td>
    </tr>

    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
  </table>

  <table cellpadding=0 cellspacing=0 border=0>
    <tr>
      <td><br></td>
    </tr>
    <tr>
      <td>
<pre class="sourceHeading">           Branch data     Line data    Source code</pre>
<pre class="source">
<a name="1"><span class="lineNum">       1 </span>           [<span class="branchCov" title="Branch 0 was taken 2 times"> + </span>]:<span class="lineCov">          2 : import assert from 'node:assert/strict';</span></a>
<a name="2"><span class="lineNum">       2 </span>                :<span class="lineCov">          2 : import { XMLHttpRequest } from '../../../lib/whatwg-xhr.js';</span></a>
<a name="3"><span class="lineNum">       3 </span>                :<span class="lineCov">          2 : </span></a>
<a name="4"><span class="lineNum">       4 </span>           [<span class="branchCov" title="Branch 0 was taken 1 time"> + </span>]:<span class="lineCov">          2 : export default (activeURL) =&gt; {</span></a>
<a name="5"><span class="lineNum">       5 </span>                :<span class="lineCov">          1 : </span></a>
<a name="6"><span class="lineNum">       6 </span>                :<span class="lineCov">          1 :         let xhr = new XMLHttpRequest();</span></a>
<a name="7"><span class="lineNum">       7 </span>                :<span class="lineCov">          1 : </span></a>
<a name="8"><span class="lineNum">       8 </span>                :<span class="lineCov">          1 :         xhr.open('POST', `${activeURL}/inspect-headers.py?filter_value=TEST`, false);</span></a>
<a name="9"><span class="lineNum">       9 </span>                :<span class="lineCov">          1 : </span></a>
<a name="10"><span class="lineNum">      10 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Accept-Charset', 'TEST');</span></a>
<a name="11"><span class="lineNum">      11 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Accept-Encoding', 'TEST');</span></a>
<a name="12"><span class="lineNum">      12 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Connection', 'TEST');</span></a>
<a name="13"><span class="lineNum">      13 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Content-Length', 'TEST');</span></a>
<a name="14"><span class="lineNum">      14 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Cookie', 'TEST');</span></a>
<a name="15"><span class="lineNum">      15 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Cookie2', 'TEST');</span></a>
<a name="16"><span class="lineNum">      16 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Date', 'TEST');</span></a>
<a name="17"><span class="lineNum">      17 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('DNT', 'TEST');</span></a>
<a name="18"><span class="lineNum">      18 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Expect', 'TEST');</span></a>
<a name="19"><span class="lineNum">      19 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Host', 'TEST');</span></a>
<a name="20"><span class="lineNum">      20 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Keep-Alive', 'TEST');</span></a>
<a name="21"><span class="lineNum">      21 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Referer', 'TEST');</span></a>
<a name="22"><span class="lineNum">      22 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('TE', 'TEST');</span></a>
<a name="23"><span class="lineNum">      23 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Trailer', 'TEST');</span></a>
<a name="24"><span class="lineNum">      24 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Transfer-Encoding', 'TEST');</span></a>
<a name="25"><span class="lineNum">      25 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Upgrade', 'TEST');</span></a>
<a name="26"><span class="lineNum">      26 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Via', 'TEST');</span></a>
<a name="27"><span class="lineNum">      27 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Proxy-', 'TEST');</span></a>
<a name="28"><span class="lineNum">      28 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Proxy-LIES', 'TEST');</span></a>
<a name="29"><span class="lineNum">      29 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Proxy-Authorization', 'TEST');</span></a>
<a name="30"><span class="lineNum">      30 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Sec-', 'TEST');</span></a>
<a name="31"><span class="lineNum">      31 </span>                :<span class="lineCov">          1 :         xhr.setRequestHeader('Sec-X', 'TEST');</span></a>
<a name="32"><span class="lineNum">      32 </span>                :<span class="lineCov">          1 : </span></a>
<a name="33"><span class="lineNum">      33 </span>                :<span class="lineCov">          1 :         xhr.send();</span></a>
<a name="34"><span class="lineNum">      34 </span>                :<span class="lineCov">          1 :         assert.strictEqual(xhr.responseText, '');</span></a>
<a name="35"><span class="lineNum">      35 </span>                :<span class="lineCov">          1 : </span></a>
<a name="36"><span class="lineNum">      36 </span>                :<span class="lineCov">          1 :         let forbiddenMethods = [</span></a>
<a name="37"><span class="lineNum">      37 </span>                :<span class="lineCov">          1 :                 'TRACE',</span></a>
<a name="38"><span class="lineNum">      38 </span>                :<span class="lineCov">          1 :                 'TRACK',</span></a>
<a name="39"><span class="lineNum">      39 </span>                :<span class="lineCov">          1 :                 'CONNECT',</span></a>
<a name="40"><span class="lineNum">      40 </span>                :<span class="lineCov">          1 :                 'trace',</span></a>
<a name="41"><span class="lineNum">      41 </span>                :<span class="lineCov">          1 :                 'track',</span></a>
<a name="42"><span class="lineNum">      42 </span>                :<span class="lineCov">          1 :                 'connect',</span></a>
<a name="43"><span class="lineNum">      43 </span>                :<span class="lineCov">          1 :                 'trace,',</span></a>
<a name="44"><span class="lineNum">      44 </span>                :<span class="lineCov">          1 :                 'GET,track ',</span></a>
<a name="45"><span class="lineNum">      45 </span>                :<span class="lineCov">          1 :                 ' connect',</span></a>
<a name="46"><span class="lineNum">      46 </span>                :<span class="lineCov">          1 :         ];</span></a>
<a name="47"><span class="lineNum">      47 </span>                :<span class="lineCov">          1 : </span></a>
<a name="48"><span class="lineNum">      48 </span>                :<span class="lineCov">          1 :         let overrideHeaders = [</span></a>
<a name="49"><span class="lineNum">      49 </span>                :<span class="lineCov">          1 :                 'x-http-method-override',</span></a>
<a name="50"><span class="lineNum">      50 </span>                :<span class="lineCov">          1 :                 'x-http-method',</span></a>
<a name="51"><span class="lineNum">      51 </span>                :<span class="lineCov">          1 :                 'x-method-override',</span></a>
<a name="52"><span class="lineNum">      52 </span>                :<span class="lineCov">          1 :                 'X-HTTP-METHOD-OVERRIDE',</span></a>
<a name="53"><span class="lineNum">      53 </span>                :<span class="lineCov">          1 :                 'X-HTTP-METHOD',</span></a>
<a name="54"><span class="lineNum">      54 </span>                :<span class="lineCov">          1 :                 'X-METHOD-OVERRIDE',</span></a>
<a name="55"><span class="lineNum">      55 </span>                :<span class="lineCov">          1 :         ];</span></a>
<a name="56"><span class="lineNum">      56 </span>                :<span class="lineCov">          1 : </span></a>
<a name="57"><span class="lineNum">      57 </span>           [<span class="branchCov" title="Branch 0 was taken 9 times"> + </span>]:<span class="lineCov">          1 :         for (let forbiddenMethod of forbiddenMethods) {</span></a>
<a name="58"><span class="lineNum">      58 </span>           [<span class="branchCov" title="Branch 0 was taken 54 times"> + </span>]:<span class="lineCov">          9 :                 for (let overrideHeader of overrideHeaders) {</span></a>
<a name="59"><span class="lineNum">      59 </span>                :<span class="lineCov">         54 :                         let xhr = new XMLHttpRequest();</span></a>
<a name="60"><span class="lineNum">      60 </span>                :<span class="lineCov">         54 : </span></a>
<a name="61"><span class="lineNum">      61 </span>                :<span class="lineCov">         54 :                         xhr.open('POST',</span></a>
<a name="62"><span class="lineNum">      62 </span>                :<span class="lineCov">         54 :                                 `${activeURL}/inspect-headers.py?filter_name=${forbiddenMethod}`, false);</span></a>
<a name="63"><span class="lineNum">      63 </span>                :<span class="lineCov">         54 : </span></a>
<a name="64"><span class="lineNum">      64 </span>                :<span class="lineCov">         54 :                         xhr.setRequestHeader(overrideHeader, forbiddenMethod);</span></a>
<a name="65"><span class="lineNum">      65 </span>                :<span class="lineCov">         54 :                         xhr.send();</span></a>
<a name="66"><span class="lineNum">      66 </span>                :<span class="lineCov">         54 : </span></a>
<a name="67"><span class="lineNum">      67 </span>                :<span class="lineCov">         54 :                         assert.strictEqual(xhr.responseText, '');</span></a>
<a name="68"><span class="lineNum">      68 </span>                :<span class="lineCov">         54 :                 }</span></a>
<a name="69"><span class="lineNum">      69 </span>                :<span class="lineCov">          9 :         }</span></a>
<a name="70"><span class="lineNum">      70 </span>                :<span class="lineCov">          1 : </span></a>
<a name="71"><span class="lineNum">      71 </span>                :<span class="lineCov">          1 :         let permittedValues = [</span></a>
<a name="72"><span class="lineNum">      72 </span>                :<span class="lineCov">          1 :                 'GETTRACE',</span></a>
<a name="73"><span class="lineNum">      73 </span>                :<span class="lineCov">          1 :                 'GET',</span></a>
<a name="74"><span class="lineNum">      74 </span>                :<span class="lineCov">          1 :                 '\&quot;,TRACE\&quot;,',</span></a>
<a name="75"><span class="lineNum">      75 </span>                :<span class="lineCov">          1 :         ];</span></a>
<a name="76"><span class="lineNum">      76 </span>                :<span class="lineCov">          1 : </span></a>
<a name="77"><span class="lineNum">      77 </span>           [<span class="branchCov" title="Branch 0 was taken 3 times"> + </span>]:<span class="lineCov">          1 :         for (let permittedValue of permittedValues) {</span></a>
<a name="78"><span class="lineNum">      78 </span>           [<span class="branchCov" title="Branch 0 was taken 18 times"> + </span>]:<span class="lineCov">          3 :                  for (let overrideHeader of overrideHeaders) {</span></a>
<a name="79"><span class="lineNum">      79 </span>                :<span class="lineCov">         18 :                         let xhr = new XMLHttpRequest();</span></a>
<a name="80"><span class="lineNum">      80 </span>                :<span class="lineCov">         18 : </span></a>
<a name="81"><span class="lineNum">      81 </span>                :<span class="lineCov">         18 :                         xhr.open('POST',</span></a>
<a name="82"><span class="lineNum">      82 </span>                :<span class="lineCov">         18 :                                 `${activeURL}/inspect-headers.py?filter_name=${overrideHeader}`, false);</span></a>
<a name="83"><span class="lineNum">      83 </span>                :<span class="lineCov">         18 : </span></a>
<a name="84"><span class="lineNum">      84 </span>                :<span class="lineCov">         18 :                         xhr.setRequestHeader(overrideHeader, permittedValue);</span></a>
<a name="85"><span class="lineNum">      85 </span>                :<span class="lineCov">         18 :                         xhr.send();</span></a>
<a name="86"><span class="lineNum">      86 </span>                :<span class="lineCov">         18 : </span></a>
<a name="87"><span class="lineNum">      87 </span>                :<span class="lineCov">         18 :                         assert.strictEqual(xhr.responseText, `${overrideHeader}: ${permittedValue}\n`);</span></a>
<a name="88"><span class="lineNum">      88 </span>                :<span class="lineCov">         18 :                 }</span></a>
<a name="89"><span class="lineNum">      89 </span>                :<span class="lineCov">          3 :         }</span></a>
<a name="90"><span class="lineNum">      90 </span>                :<span class="lineCov">          1 : }</span></a>
<a name="91"><span class="lineNum">      91 </span>                :<span class="lineCov">          2 : </span></a>
<a name="92"><span class="lineNum">      92 </span>                :<span class="lineCov">          2 : /*</span></a>
<a name="93"><span class="lineNum">      93 </span>                :<span class="lineCov">          2 :  * setrequestheader-header-forbidden.htm</span></a>
<a name="94"><span class="lineNum">      94 </span>                :<span class="lineCov">          2 :  *</span></a>
<a name="95"><span class="lineNum">      95 </span>                :<span class="lineCov">          2 : </span></a>
<a name="96"><span class="lineNum">      96 </span>                :<span class="lineCov">          2 : &lt;!doctype html&gt;</span></a>
<a name="97"><span class="lineNum">      97 </span>                :<span class="lineCov">          2 : &lt;html&gt;</span></a>
<a name="98"><span class="lineNum">      98 </span>                :<span class="lineCov">          2 :   &lt;head&gt;</span></a>
<a name="99"><span class="lineNum">      99 </span>                :<span class="lineCov">          2 :     &lt;title&gt;XMLHttpRequest: setRequestHeader() - headers that are forbidden&lt;/title&gt;</span></a>
<a name="100"><span class="lineNum">     100 </span>                :<span class="lineCov">          2 :     &lt;script src=&quot;/resources/testharness.js&quot;&gt;&lt;/script&gt;</span></a>
<a name="101"><span class="lineNum">     101 </span>                :<span class="lineCov">          2 :     &lt;script src=&quot;/resources/testharnessreport.js&quot;&gt;&lt;/script&gt;</span></a>
<a name="102"><span class="lineNum">     102 </span>                :<span class="lineCov">          2 :     &lt;link rel=&quot;help&quot; href=&quot;https://xhr.spec.whatwg.org/#the-setrequestheader()-method&quot;&gt;</span></a>
<a name="103"><span class="lineNum">     103 </span>                :<span class="lineCov">          2 : </span></a>
<a name="104"><span class="lineNum">     104 </span>                :<span class="lineCov">          2 :   &lt;/head&gt;</span></a>
<a name="105"><span class="lineNum">     105 </span>                :<span class="lineCov">          2 :   &lt;body&gt;</span></a>
<a name="106"><span class="lineNum">     106 </span>                :<span class="lineCov">          2 :     &lt;div id=&quot;log&quot;&gt;&lt;/div&gt;</span></a>
<a name="107"><span class="lineNum">     107 </span>                :<span class="lineCov">          2 :     &lt;script&gt;</span></a>
<a name="108"><span class="lineNum">     108 </span>                :<span class="lineCov">          2 :       test(function() {</span></a>
<a name="109"><span class="lineNum">     109 </span>                :<span class="lineCov">          2 :         var client = new XMLHttpRequest()</span></a>
<a name="110"><span class="lineNum">     110 </span>                :<span class="lineCov">          2 :         client.open(&quot;POST&quot;, &quot;resources/inspect-headers.py?filter_value=TEST&quot;, false)</span></a>
<a name="111"><span class="lineNum">     111 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Accept-Charset&quot;, &quot;TEST&quot;)</span></a>
<a name="112"><span class="lineNum">     112 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Accept-Encoding&quot;, &quot;TEST&quot;)</span></a>
<a name="113"><span class="lineNum">     113 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Connection&quot;, &quot;TEST&quot;)</span></a>
<a name="114"><span class="lineNum">     114 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Content-Length&quot;, &quot;TEST&quot;)</span></a>
<a name="115"><span class="lineNum">     115 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Cookie&quot;, &quot;TEST&quot;)</span></a>
<a name="116"><span class="lineNum">     116 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Cookie2&quot;, &quot;TEST&quot;)</span></a>
<a name="117"><span class="lineNum">     117 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Date&quot;, &quot;TEST&quot;)</span></a>
<a name="118"><span class="lineNum">     118 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;DNT&quot;, &quot;TEST&quot;)</span></a>
<a name="119"><span class="lineNum">     119 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Expect&quot;, &quot;TEST&quot;)</span></a>
<a name="120"><span class="lineNum">     120 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Host&quot;, &quot;TEST&quot;)</span></a>
<a name="121"><span class="lineNum">     121 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Keep-Alive&quot;, &quot;TEST&quot;)</span></a>
<a name="122"><span class="lineNum">     122 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Referer&quot;, &quot;TEST&quot;)</span></a>
<a name="123"><span class="lineNum">     123 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;TE&quot;, &quot;TEST&quot;)</span></a>
<a name="124"><span class="lineNum">     124 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Trailer&quot;, &quot;TEST&quot;)</span></a>
<a name="125"><span class="lineNum">     125 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Transfer-Encoding&quot;, &quot;TEST&quot;)</span></a>
<a name="126"><span class="lineNum">     126 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Upgrade&quot;, &quot;TEST&quot;)</span></a>
<a name="127"><span class="lineNum">     127 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Via&quot;, &quot;TEST&quot;)</span></a>
<a name="128"><span class="lineNum">     128 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Proxy-&quot;, &quot;TEST&quot;)</span></a>
<a name="129"><span class="lineNum">     129 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Proxy-LIES&quot;, &quot;TEST&quot;)</span></a>
<a name="130"><span class="lineNum">     130 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Proxy-Authorization&quot;, &quot;TEST&quot;)</span></a>
<a name="131"><span class="lineNum">     131 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Sec-&quot;, &quot;TEST&quot;)</span></a>
<a name="132"><span class="lineNum">     132 </span>                :<span class="lineCov">          2 :         client.setRequestHeader(&quot;Sec-X&quot;, &quot;TEST&quot;)</span></a>
<a name="133"><span class="lineNum">     133 </span>                :<span class="lineCov">          2 :         client.send(null)</span></a>
<a name="134"><span class="lineNum">     134 </span>                :<span class="lineCov">          2 :         assert_equals(client.responseText, &quot;&quot;)</span></a>
<a name="135"><span class="lineNum">     135 </span>                :<span class="lineCov">          2 :         })</span></a>
<a name="136"><span class="lineNum">     136 </span>                :<span class="lineCov">          2 : </span></a>
<a name="137"><span class="lineNum">     137 </span>                :<span class="lineCov">          2 :         test (function() {</span></a>
<a name="138"><span class="lineNum">     138 </span>                :<span class="lineCov">          2 : </span></a>
<a name="139"><span class="lineNum">     139 </span>                :<span class="lineCov">          2 :         let forbiddenMethods = [</span></a>
<a name="140"><span class="lineNum">     140 </span>                :<span class="lineCov">          2 :           &quot;TRACE&quot;,</span></a>
<a name="141"><span class="lineNum">     141 </span>                :<span class="lineCov">          2 :           &quot;TRACK&quot;,</span></a>
<a name="142"><span class="lineNum">     142 </span>                :<span class="lineCov">          2 :           &quot;CONNECT&quot;,</span></a>
<a name="143"><span class="lineNum">     143 </span>                :<span class="lineCov">          2 :           &quot;trace&quot;,</span></a>
<a name="144"><span class="lineNum">     144 </span>                :<span class="lineCov">          2 :           &quot;track&quot;,</span></a>
<a name="145"><span class="lineNum">     145 </span>                :<span class="lineCov">          2 :           &quot;connect&quot;,</span></a>
<a name="146"><span class="lineNum">     146 </span>                :<span class="lineCov">          2 :           &quot;trace,&quot;,</span></a>
<a name="147"><span class="lineNum">     147 </span>                :<span class="lineCov">          2 :           &quot;GET,track &quot;,</span></a>
<a name="148"><span class="lineNum">     148 </span>                :<span class="lineCov">          2 :           &quot; connect&quot;,</span></a>
<a name="149"><span class="lineNum">     149 </span>                :<span class="lineCov">          2 :         ];</span></a>
<a name="150"><span class="lineNum">     150 </span>                :<span class="lineCov">          2 : </span></a>
<a name="151"><span class="lineNum">     151 </span>                :<span class="lineCov">          2 :         let overrideHeaders = [</span></a>
<a name="152"><span class="lineNum">     152 </span>                :<span class="lineCov">          2 :           &quot;x-http-method-override&quot;,</span></a>
<a name="153"><span class="lineNum">     153 </span>                :<span class="lineCov">          2 :           &quot;x-http-method&quot;,</span></a>
<a name="154"><span class="lineNum">     154 </span>                :<span class="lineCov">          2 :           &quot;x-method-override&quot;,</span></a>
<a name="155"><span class="lineNum">     155 </span>                :<span class="lineCov">          2 :           &quot;X-HTTP-METHOD-OVERRIDE&quot;,</span></a>
<a name="156"><span class="lineNum">     156 </span>                :<span class="lineCov">          2 :           &quot;X-HTTP-METHOD&quot;,</span></a>
<a name="157"><span class="lineNum">     157 </span>                :<span class="lineCov">          2 :           &quot;X-METHOD-OVERRIDE&quot;,</span></a>
<a name="158"><span class="lineNum">     158 </span>                :<span class="lineCov">          2 :         ];</span></a>
<a name="159"><span class="lineNum">     159 </span>                :<span class="lineCov">          2 : </span></a>
<a name="160"><span class="lineNum">     160 </span>                :<span class="lineCov">          2 :         for (forbiddenMethod of forbiddenMethods) {</span></a>
<a name="161"><span class="lineNum">     161 </span>                :<span class="lineCov">          2 :           for (overrideHeader of overrideHeaders) {</span></a>
<a name="162"><span class="lineNum">     162 </span>                :<span class="lineCov">          2 :              var client = new XMLHttpRequest()</span></a>
<a name="163"><span class="lineNum">     163 </span>                :<span class="lineCov">          2 :              client.open(&quot;POST&quot;,</span></a>
<a name="164"><span class="lineNum">     164 </span>                :<span class="lineCov">          2 :                      `resources/inspect-headers.py?filter_value=${forbiddenMethod}`, false)</span></a>
<a name="165"><span class="lineNum">     165 </span>                :<span class="lineCov">          2 :              client.setRequestHeader(overrideHeader, forbiddenMethod)</span></a>
<a name="166"><span class="lineNum">     166 </span>                :<span class="lineCov">          2 :              client.send(null)</span></a>
<a name="167"><span class="lineNum">     167 </span>                :<span class="lineCov">          2 :              assert_equals(client.responseText, &quot;&quot;)</span></a>
<a name="168"><span class="lineNum">     168 </span>                :<span class="lineCov">          2 :           }</span></a>
<a name="169"><span class="lineNum">     169 </span>                :<span class="lineCov">          2 :         }</span></a>
<a name="170"><span class="lineNum">     170 </span>                :<span class="lineCov">          2 : </span></a>
<a name="171"><span class="lineNum">     171 </span>                :<span class="lineCov">          2 :         let permittedValues = [</span></a>
<a name="172"><span class="lineNum">     172 </span>                :<span class="lineCov">          2 :         &quot;GETTRACE&quot;,</span></a>
<a name="173"><span class="lineNum">     173 </span>                :<span class="lineCov">          2 :         &quot;GET&quot;,</span></a>
<a name="174"><span class="lineNum">     174 </span>                :<span class="lineCov">          2 :         &quot;\&quot;,TRACE\&quot;,&quot;,</span></a>
<a name="175"><span class="lineNum">     175 </span>                :<span class="lineCov">          2 :         ];</span></a>
<a name="176"><span class="lineNum">     176 </span>                :<span class="lineCov">          2 : </span></a>
<a name="177"><span class="lineNum">     177 </span>                :<span class="lineCov">          2 :         for (permittedValue of permittedValues) {</span></a>
<a name="178"><span class="lineNum">     178 </span>                :<span class="lineCov">          2 :           for (overrideHeader of overrideHeaders) {</span></a>
<a name="179"><span class="lineNum">     179 </span>                :<span class="lineCov">          2 :              var client = new XMLHttpRequest()</span></a>
<a name="180"><span class="lineNum">     180 </span>                :<span class="lineCov">          2 :              client.open(&quot;POST&quot;,</span></a>
<a name="181"><span class="lineNum">     181 </span>                :<span class="lineCov">          2 :                      `resources/inspect-headers.py?filter_name=${overrideHeader}`, false)</span></a>
<a name="182"><span class="lineNum">     182 </span>                :<span class="lineCov">          2 :              client.setRequestHeader(overrideHeader, permittedValue)</span></a>
<a name="183"><span class="lineNum">     183 </span>                :<span class="lineCov">          2 :              client.send(null)</span></a>
<a name="184"><span class="lineNum">     184 </span>                :<span class="lineCov">          2 :              assert_equals(client.responseText, overrideHeader + &quot;: &quot; + permittedValue + &quot;\n&quot;)</span></a>
<a name="185"><span class="lineNum">     185 </span>                :<span class="lineCov">          2 :           }</span></a>
<a name="186"><span class="lineNum">     186 </span>                :<span class="lineCov">          2 :         }</span></a>
<a name="187"><span class="lineNum">     187 </span>                :<span class="lineCov">          2 :       })</span></a>
<a name="188"><span class="lineNum">     188 </span>                :<span class="lineCov">          2 :     &lt;/script&gt;</span></a>
<a name="189"><span class="lineNum">     189 </span>                :<span class="lineCov">          2 :   &lt;/body&gt;</span></a>
<a name="190"><span class="lineNum">     190 </span>                :<span class="lineCov">          2 : &lt;/html&gt;</span></a>
<a name="191"><span class="lineNum">     191 </span>                :<span class="lineCov">          2 : </span></a>
<a name="192"><span class="lineNum">     192 </span>                :<span class="lineCov">          2 :  *</span></a>
<a name="193"><span class="lineNum">     193 </span>                :<span class="lineCov">          2 :  * setrequestheader-header-forbidden.htm</span></a>
<a name="194"><span class="lineNum">     194 </span>                :<span class="lineCov">          2 :  */</span></a>
</pre>
      </td>
    </tr>
  </table>
  <br>

  <table width="100%" border=0 cellspacing=0 cellpadding=0>
    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
    <tr><td class="versionInfo">Generated by: <a href="http://ltp.sourceforge.net/coverage/lcov.php" target="_parent">LCOV version 1.14</a></td></tr>
  </table>
  <br>

</body>
</html>
